Add implementation of PLACE command
With the bare bones of the Robot and the Board in place, the PLACE
command is the first command to be implemented, since all other
commands rely on this being called first by the user.

The other commands could be added independantly but wouldn't give me the
insight or tests through board that I would like.

Direction has been omitted for simplicity in this initial place method
but will be added next.