Implement MOVE command
This implements MOVE by adjusting the position
based on the direction faced. Validation handled by the
board before the move command send, hence no test for
invalid moves in the robot spec, as that class is not
concerned with validity.

Would prefer for some of this to be split into respective
classes but would require a bigger refactor due to choices
already made.